cloud_functions_checking_motion/requirements.txt
functions-framework==3.*
google-cloud-iot
google-cloud-pubsub
google-cloud-bigquery
google-cloud-aiplatform